+++ lib/src/ordering/nulls_first.dart
@@ -19,4 +19,13 @@ class NullsFirstOrdering<T> extends Ordering<T> {
       return ordering.compare(a, b);
     }
   }
+
+  @override
+  Ordering<T> get reversed => ordering.reversed.nullsLast;
+
+  @override
+  Ordering<T> get nullsFirst => this;
+
+  @override
+  Ordering<T> get nullsLast => ordering.nullsLast;
 }